refactor(server): extract template builders for reuse
