+++ src/server/mod.rs
@@ -49,6 +49,7 @@ use std::{net::IpAddr, net::SocketAddr, sync::Arc};
 use tower_http::{cors::CorsLayer, services::ServeDir};
 use tracing::{error, info};
 
+pub mod builders;
 mod handlers;
 mod i18n;
 mod language;
@@ -57,7 +58,7 @@ mod shopping_list_store;
 mod shopping_list_watcher;
 #[cfg(feature = "sync")]
 pub mod sync;
-mod templates;
+pub mod templates;
 mod ui;
 
 // Embed static files at compile time
